Make indentation handling easier to follow
This commits adds a new function and a context manager to the python
code.

The new function generates some output with one extra level of
indentation. This function allows to skip changes to the indentation
level for single lines.

The context manager increases indentation while it is active. Thus,
indentation in the generated code corresponds with indentation in the
code generator. Hopefully, this improves readability.

Signed-off-by: Uli Schlachter <psychon@znc.in>