Add proper support for ListFontsWithInfo
ListFontsWithInfo is the only X11 request (that I know about...) that
generates multiple replies. As such, it requires some special handling.

This commit adds a special cookie type for this request that implements
Iterator. Via this, one can access the various replies.

To showcase and test this new request, a new example program is added.
This one should produce output identical to `xlsfonts -lu`.

Signed-off-by: Uli Schlachter <psychon@znc.in>