Do not use CSlice were it makes no sense
This code used CSlice and std::mem::forget() because it was not supposed
to free the memory. However, if something in here panics, the memory
would still be freed.

It makes more sense to just use std::slice::from_raw_parts() directly
here, because that does exactly what is needed.

Signed-off-by: Uli Schlachter <psychon@znc.in>