Add one more deny: single_use_lifetimes
At first this one seemed complicated, but then the fix was indeed rather
easy: In requests with replies, the cookie borrows the connection and
this needs to be made explicit if there are other reference parameters.
However, in requests without replies, there is no cookie and thus no
explicit lifetime is needed.

Signed-off-by: Uli Schlachter <psychon@znc.in>