Add support for extensions
This commit adds basic support for extensions. All extensions which
cause errors in either the code generator or the compiler where
disabled.

With this commit, code for 11 extensions is generated while 23
extensions are currently unsupported.

This commit only provides the necessary infrastructure. The
implementation in the connections is currently unimplemented!().

Signed-off-by: Uli Schlachter <psychon@znc.in>