Skip FDs in the code generator
FD passing is not implemented yet. Instead of erroring our when
encountering <fd> fields, this commit makes the code generator emit
implementations saying "unimplemented!()". This means that at least the
other parts of the extensions in questions can be used.

This makes the DRI3 and SHM extensions "work". PRESENT does not work
yet, because it depends on SYNC which does not work for other reasons.

Signed-off-by: Uli Schlachter <psychon@znc.in>