Remove unnecessary double into(into()) from generated code
I think that when I wrote this code, enums only implemented Into for the
smallest size that contains them. Since a while now, enums can also be
converted into larger sized integers, so this double Into is no longer
necessary.

This might or might not fix the Travis build which for unknown reasons
currently fails in the code generator around this code.

(I guess the Travis failure is because the code actually depends on the
order that the XML files are read from the file system. If xproto.xml
comes first, then enum_sizes["CW"] is already defined when
screensaver.xml is parsed. Otherwise, well, KeyError.)

Signed-off-by: Uli Schlachter <psychon@znc.in>