Give the length field the correct type
All other fields get their type from the argument or from a constant.
However, the length field is computed and for $REASONS it is computed
with type usize. But on the wire, the length field has just 16 bits.

This commit ensures that we get a failure from TryInto if the length
field overflows.

Signed-off-by: Uli Schlachter <psychon@znc.in>