Add a tutorial
This tutorial was created by tacking the libxcb tutorial and translating
it to x11rb. Instead of providing a HTML page, this tutorial is written
as a rust file with the actual tutorial in comments and the code
examples as, well, code. This way, this code is checked to make sure it
actually compiles.

Signed-off-by: Uli Schlachter <psychon@znc.in>