Fix padding calculation
The padding calculation was using the length of the list that was passed
in. This works for &[u8], but does the wrong thing for e.g. &[u32],
because for each list item there are actually four bytes on the wire.
Fix this by instead using the length of the actual on-wire length.

Fixes: https://github.com/psychon/x11rb/issues/19
Signed-off-by: Uli Schlachter <psychon@znc.in>