Clarify the use of unsafe.
I was disappointed that xcb-rs requires uses of `unsafe` in code using
the library (and the `unsafe` basically is a direct call to
`std::mem::transmute`, which is quite high on "the scale of unsafety").
This lead to me wording this paragraph poorly.

Improve this by not claiming that there is no unsafe code and by
actually explaining what uses of unsafe there are.

Thanks to "po8" for pointing this out on Reddit.

Signed-off-by: Uli Schlachter <psychon@znc.in>