Fix a couple of length calculations
Before this commit, the code generator plainly ignored expressions on
<list> elements. It turns out that there is actually a reason why these
exists.

This commit makes the parsing code actually use these expressions to get
the length of the list.

The request sending code is not so lucky and instead the affected
functions gain an extra argument for the length field. Dealing with this
extra length argument is left as future work.

Affected functions:

- change_property() gains a data_len argument
- change_keyboard_mapping() gains a keycode_count argument

I carefully looked at the changes this commit causes for the generated
code for xproto.rs. There is lots of noise due to a new `let list_length
= expr;` that does not actually do much for most lists, because the
expression on the right hand is just a simple variable.

For some things, this change actually fixes bugs:

- SetupAuthenticate now correctly uses length*4 instead of just length.
- GetPropertyReply now correctly uses value_len*format/8 instead of just
  value_len. However, GetPropertyReply now also has a new value_len
  member.
- GetImageReply now correctly uses length*4 instead of just length.
- GetModifierMappingReply now correctly uses keycodes_per_modifier*8
  instead of just keycodes_per_modifier.

Signed-off-by: Uli Schlachter <psychon@znc.in>