Add the missing empty slices for xcb_send_request()
xcb_send_request() gets an array of slices that should be send to the
X11 server. For internal management reasons, it might need to access the
entries -1 and -2 of this array. This commit adds these entries so that
libxcb does not stomp on some random other piece of memory.

If someone knows how to make this code nicer, feel free to speak up.

Fixes: https://github.com/psychon/x11rb/issues/6
Signed-off-by: Uli Schlachter <psychon@znc.in>