Add pycodestyle checking to .travis.yml
The maximum line length is set to what is currently the longest line in
the Python code. While this is way too large, at least it makes sure
that nothing even larger will get merged.

The idea is that I improve line length and then also remember to update
the limit here. Let's see how well that works out...

Signed-off-by: Uli Schlachter <psychon@znc.in>