Handle <switch> with different sized cases
Sync's CreateAlarm request contains a <switch> that has both 4 byte
values and 8 byte values. The code for serialising such a switch already
works correctly by accident (it just calls to_ne_bytes() on the inner
value).

However, the number of bytes on the wire needs some special care.
Previously, the size of the values inside the <switch> was recorded and
multiplied by the number of fields that are actually present. This
request instead adds a wire_length() method on the aux struct that
provides the necessary information.

This wire_length() method is intentionally not public, because it is
ugly and I hope to get rid of it in the future again.

Signed-off-by: Uli Schlachter <psychon@znc.in>