Reduce use of use foo::*;
There is an <enum name="Transform"> in randr.xml. There is also
<import>render</import>. And render has a <struct name="TRANSFORM">. The
result of this is that there are two types in scope which both have the
name "Transform" and the rust compiler complains.

Fix this by not importing all the contents of a module. Instead, we
explicitly refer to members of the module where needed.

Signed-off-by: Uli Schlachter <psychon@znc.in>