Assert list length in requests
A list is complex if we cannot just derive its length value from a given
slice. For example, InternAtom has a simple length field, because the
length explicitly appears in a value that is sent to the server.
ChangeKeyboardMapping has a complex list length, because its length is
keycode_count * keysyms_per_keycode. Thus, the caller has to provide
both numbers explicitly.

This commit adds an assert that ensures that the caller provided a slice
of the expected length for a complex list.

Fixes: https://github.com/psychon/x11rb/issues/33
Signed-off-by: Uli Schlachter <psychon@znc.in>