Make XCBConnection Send&Sync
The only reason why this is neither Send nor Sync is that it contains a
raw pointer to the underlying xcb_connection_t. Since libxcb is
(supposedly) fully thread-safe, this is fine. Thus, this commits adds a
container struct around xcb_connection_t and declares that one to be
Send and Sync.

Signed-off-by: Uli Schlachter <psychon@znc.in>