Add support for BigRequests
This commit adds support for the big requests extension. The generated
code continues to only supports requests with a length field that fits
into an u16. If the length field overflows, it is instead set to zero.
The connection then has to handle this and insert the correct length
field.

The actual work of doing so is done by the provided method
Connection::compute_length_field(), so that this work does not have to
be duplicated by every implementation of Connection.

Fixes: https://github.com/psychon/x11rb/issues/8
Signed-off-by: Uli Schlachter <psychon@znc.in>