Factor out one function in the code generator
Instead of having an argument in complex_type() that is set to False to
disable some feature, the code that generates a TryFrom implementation
is moved to a new helper function. This way, only the callers that want
this have to deal with it.

Besides changing the order of some generated code, this has no effects
on the generated code.

Signed-off-by: Uli Schlachter <psychon@znc.in>