Add more wire fields to parsed structs
This commit causes less fields to be ignored in parsing. Specifically,
only length fields and padding is ignored. The value of length fields
can later be inferred from the length of the corresponding list. Padding
does not contain any useful data.

This means that events now also contain the response_type and sequence
fields. While the response_type for an event is not helpful (it is
always the same value for a given event), the sequence can be useful.
Well, and the response type can be useful if I ever implement
serialisation of events for SendEvent.

Errors now contain the response_type, error_code and sequence fields.
The first two are not that useful (response_type is always 0 and
error_code is again known beforehand), but sequence can be useful.

Last, but not least, replies now have a response_type, sequence, and
length fields.

Signed-off-by: Uli Schlachter <psychon@znc.in>