Turn the only_if_exists argument of InternAtom into bool
This is a bit of a hack. In theory, the XML contains the information
that this type is a boolean, but xcbgen turns this into uint8_t before
we can do anything and thus this looks the same as CARD8 to us.

The work around here is to add a special case for InternAtom to the code
generator that replaces the argument with a boolean.

There are lots of other BOOLs in the XML, but I think that InternAtom is
the most important one, so for now I will ignore all the other fields.

Fixes: https://github.com/psychon/x11rb/issues/5
Signed-off-by: Uli Schlachter <psychon@znc.in>