Change the API to allow sending of requests
This commit adds a list of file descriptors argument to the functions
for sending requests. The code generator is patched to send an empty
slice and the implementations of Connection are patched to assert that
this slice is empty.

Thus, this commit only changes the API, but does not make anything
actually work in FD sending.

Related-to: https://github.com/psychon/x11rb/issues/2
Signed-off-by: Uli Schlachter <psychon@znc.in>