Add and use a RawFdContainer type
This type represents an owned RawFd. When it is dropped, the underlying
fd is closed.

This type is used for fd passing to make it clear that the connection
takes ownership of the fd.

To make this even clearer, an API change is needed: Instead of taking a
&[RawFd] argument, the send_request_* functions now take a
Vec<RawFdContainer> argument. This is the part that makes it clear that
ownership of the FDs is transferred.

Signed-off-by: Uli Schlachter <psychon@znc.in>