Fix crash in xcb with broken connections
When an xcb connection is in an error state (meaning that
xcb_connection_has_error() returns something else than 0), it is
possible for xcb_wait_for_reply64() to return both NULL reply and error.
Before this commit, the code did not handle this case well and
dereferenced a NULL pointer. After this patch, an Err is returned
instead.

Signed-off-by: Uli Schlachter <psychon@znc.in>