Implement XCBConnection::wait_for_reply_with_fds()
This request gets rid of the unimplemented!() in
XCBConnection::wait_for_reply_with_fds(). xcb simply saves a list of
integers behind the response in memory. This list contains the file
descriptors that were received.

Thus, the code does exactly that: It gets the number of FDs from the
reply and then reads the list of FDs into a Vec<RawFdContainer>. The
result is given to the caller (where the support for handling this was
added in the previous two commits).

This is the final piece in implementing
https://github.com/psychon/x11rb/issues/2.

Signed-off-by: Uli Schlachter <psychon@znc.in>