Fix name mangling for enum kinds
XInput has a ChangeDeviceProperty request that references a
PropertyFormat enum. This enum has values like "8Bits". Since this is
not a valid Rust identifier, the code generator changes this into M8Bit.
However, it generated code that later referred to "8Bits".

This commit fixes the code to refer to the enum variants with the right
name.

Signed-off-by: Uli Schlachter <psychon@znc.in>