Code generator: Only simple types can be length fields
The xinput extension has a ListInputDevices request. The reply of this
request contains:

    <list type="DeviceInfo name="devices">
        [snip]
    </list>
    <list type="InputInfo" name="infos">
        <sumof ref="devices">
            [snip]
        </sumof>
    </list>

With the current code, this caused the whole "devices" list to be marked
as the length field of the "infos" list. This is clearly incorrect.

This commit changes the code so that <sumof> values are not used as
length field names. This still caused "devices" to be skipped, because
get_references() is only used to decide if a list is "simple" or
"complex".

Thus, this commit also changes the code generator so that only "simple"
types can be length fields. This causes the list in question to be left
alone.

Signed-off-by: Uli Schlachter <psychon@znc.in>