Add support for "unresolved members"
XInput has a struct with a member that is... well, not a member:

    <struct name="DeviceTimeCoord">
        <field type="TIMESTAMP" name="time" />
        <list type="INT32" name="axisvalues">
            <paramref type="CARD8">num_axes</paramref>
        </list>
    </struct>

The num_axes parameter is not a member of the struct. Instead, the only
use of this struct, the reply to the GetDeviceMotionEvents request, has
a num_axes member. This member is used to specify the list of the
DeviceTimeCoord struct.

This weird construction is needed because the DeviceTimeCoord structs
themselves appear in a list.

Signed-off-by: Uli Schlachter <psychon@znc.in>