Add a makefile for testing code gen
This Makefile simplifies testing changes to the code generator.

With "make reference", the code generator is run and the generated code
is saved inside of target/.

With "make check", the code generator is run again and this time the
generated code is compared with the previously saved reference. This
allows to easily see the effect of changes to the code generator.

Signed-off-by: Uli Schlachter <psychon@znc.in>