Improve cookie API
This adds a function sequence_number() to all cookies that allows to get
the sequence number of the request that generated this cookie.

All cookies also get a raw_reply() method that returns the reply as a
Buffer instead of "already parsed". This might be useful for people. One
exception is the ListFontsWithInfoCookie: I failed to make raw_reply()
work here, so the cookie does not get that method.

Signed-off-by: Uli Schlachter <psychon@znc.in>