Remove Option from RawCookie
RawCookie contains an Option<SequenceNumber>. This Option is (basically)
always a Some() instance. The only exception is when the RawCookie is
consumed. Then, this Option becomes None. This indicates to drop() that
it should not call discard_reply().

This commit gets rid of the Option<>. Instead, std::mem::forget() is
used to prevent drop() from doing its thing. This gets rid of some calls
to unwrap() which "can never fail", but are still bad/annoying/ugly.

This changes causes some follow-up changes to ListFontsWithInfoCookie.
It now explicitly contains an Option<RawCookie> instead of (ab)using the
Option that was in RawCookie.

This change also uncovers that the previous code actually introduced a
possibility to cause None.unwrap() to run. Whoops.
(ListFontsWithInfo::sequence_number() after all replies were consumed)

Signed-off-by: Uli Schlachter <psychon@znc.in>