Change computation of total request length
Up to now, the total request length appeared out of nowhere in the
generated code and one had to check if everything really was included.
This commit adds a "length_so_far" variable that contains the length of
all the parts that were handled so far. After every part of the request,
"let length_so_far = length_so_far + new_part.len()" is used to update
the length.

This commit also fixes Python 2 compatibility of the code generator.
'nonlocal' is not available in Python 2. Sigh...

Signed-off-by: Uli Schlachter <psychon@znc.in>