Improve union length calculation
Union members can have lots of sizes and... it seems like they do not
have to all have the same size. At least, XKB's "Action" union has a
(fake?) "type" variant that only provides the type of the contained
value, I guess?

Anyway, this commit makes the code generator not do any assertions about
the size of union members and it converts the code to use xcbgen's
get_total_size() to calculate the size of a union.

Signed-off-by: Uli Schlachter <psychon@znc.in>