Mock libxcb for tests
This adds some implementation of the xcb functions that are used in
cfg(test). That way, we can test XCBConnection without having to
actually have a working X11 server, i.e. we can do unit tests.

(Plus, this should make it possible to run at least 'cargo test' on
Appveyor)

Signed-off-by: Uli Schlachter <psychon@znc.in>