Code generator: Add support for XGE events
XGE events can have sizes larger than 32 bytes. This commit adds the
necessary code to the code generator.

Since the constructor of x11rb::x11_utils::GenericEvent ensures that the
buffer has at least 32 bytes, "normal" events can implement
From<GenericEvent>. However, this is not the case for XGE events and
thus these events only implement TryFrom.

Also, normal events implement Into<[u8; 32>]. This is also not
implemented for XGE events.

Finally, this adds support for types with length calculate_len. These
seem to be lists that use the length field of whatever contains them and
then "everything to the end" is part of this list.

Fixes: https://github.com/psychon/x11rb/issues/11
Signed-off-by: Uli Schlachter <psychon@znc.in>