Create global functions for each request
Before this commit, every extension defined an extension trait that
contained the requests that could be sent. This commit keeps the
extension trait, but also adds again the global functions that can be
called. Thus, this partly reverts commit 6ad79697012, but only in the
"bring back the global functions"-sense, not as an actual revert.

The reason why this is done is name conflicts: For example, basically
everything comes with a QueryVersion request, so conn.query_version()
becomes ambiguous when multiple extensions trait are in scope. In this
case, one has to do e.g. shm::ConnectionExt::query_version(&conn)
instead. With this commit, one can do shm::query_version(&conn) instead,
which is slightly nicer.

Signed-off-by: Uli Schlachter <psychon@znc.in>