Return a copy from extension_information()
Before this commit, extension_information() returned a reference to the
QueryExtensionReply. However, this struct has a size of only 11 bytes
and implements Copy. Just returning a copy instead simplifies things a
bit. Most of all, it gets rid of an unsafe.

Thanks to @dalcde for pointing out that QueryExtensionReply is small and
implements Copy.

Signed-off-by: Uli Schlachter <psychon@znc.in>