Add clippy to Travis CI
This commit makes Travis run clippy and abort on warnings/errors. This
can help identify questionable patterns and to get rid of them.

To make this work, this commit adds an #[allow] to get rid of a warning
for such a questionable pattern. Future work should get rid of this
again. However, for now there is value in having clippy, even if part of
it is being silenced by this #[allow].

Thanks to @dalcde for making me do this.

Signed-off-by: Uli Schlachter <psychon@znc.in>