Convert BOOL in XML to bool in Rust
Thanks to the recent API addition in xcb-proto, namely the xml_type
property on SimpleType, it becomes possible to detect fields with type
BOOL. These are now represented as bool instead of u8 by the code
generator.

At the same time, this preserves backwards-compatibility with xcb-proto
without this xml_type property.

Signed-off-by: Uli Schlachter <psychon@znc.in>