Some type handling fixes
In the code for the SYNC extension, the built-in INT64 type (which maps
to i64) was confused with the INT64 struct that SYNC defines. This
commit fixes this confusion.

Also, this starts work on centralising all the "how to format this
name?"-handling in the _name() function. For this purpose, the function
starts looking up the type described by its argument.

Signed-off-by: Uli Schlachter <psychon@znc.in>