Use nicer type names in the generated code
For example, before this commit, destroy_window() took an argument of
type u32. After this commit, the argument has type WINDOW. These kind of
changes should improve readability everywhere, because the types now
have a meaning.

Of course, this does not actually mean that the compiler enforces
anything. WINDOW is still only a type alias for u32.

Fixes: https://github.com/psychon/x11rb/issues/135
Signed-off-by: Uli Schlachter <psychon@znc.in>