Readability: Improve bool handling
Instead of having the "magical dance" required for identifying bool
types around everywhere, this moves it to an is_bool() helper function.
The name of the function makes the intent much clearer here.

Additionally, this adds a _field_type() method that provides the type
for a field. This function just has to handle booleans specially and
then delegates to _name(field.field_type).

Signed-off-by: Uli Schlachter <psychon@znc.in>