Use get_derives() also for aux structures
XKB's SetMap request sends variable-sized lists as an optional field.
This turns into a field with type Option<Vec<Something>> on the aux
structure. Obviously, this interferes with #[derive(Copy)].

Fix this by using the existing infrastructure for getting the possible
derives.

This commit does not cause changes to any of the already working code.

Signed-off-by: Uli Schlachter <psychon@znc.in>