Remove wire_length()
This commit changes the code for emitting a request with <switch>.

Previously, the code would first call .wire_length() on the aux struct
to calculate the length of the request. Afterwards, it would call
.to_ne_bytes() to actually construct the on-the-wire representation of
the aux struct.

This commit moves the call to .to_ne_bytes() for aux structs up before
the length calculation. Then, the on-the-wire length of the <switch> is
simply .len() on the resulting Vec<>.

Thus, the effect of this commit on request functions is just to move one
line to an earlier place and change one line slightly (_bytes.len()
instead of .wire_length()).

The effect of this small change is that the wire_length() function on
aux structs becomes unused and can be removed. Thanks to this,
supporting variable-length aux structs (which occur in the XKB
extension) becomes easier. However, it is of course also nice that this
commit makes the generated code smaller.

Signed-off-by: Uli Schlachter <psychon@znc.in>