Fix some variable names
In some cases, the code generator used the names from the XML as
variable names as-is while in other it mangled them to conform to Rust
conventions. This commit fixes some places that were missing some name
mangling, causing mismatches.

This commit does not cause changes to the code for existing extension,
but it helps with XKB extension support.

Signed-off-by: Uli Schlachter <psychon@znc.in>