Convert more code to _field_type()
The pattern _name(field.type.name) is always wrong, because it generates
types like 'u32' even though 'ATOM' might be better. This commit changes
all the instances of this pattern that I could find to use the correct
_field_type(field) instead. However, these instances are all in places
where it does not matter (mostly length fields, which are always numbers
and never things like ATOM).

Signed-off-by: Uli Schlachter <psychon@znc.in>