Add & use a helper for computing owned types
A list of unknown length is a slice ([T]) when the user provides it to
us, but a vector (Vec<T>) when we have to own it. For the first case,
there was already a helper function _to_complex_rust_type(). This commit
adds a function _to_complex_owned_rust_type(). This reduces some code
duplication, plus this commit also fixes a case where the old code did
the wrong thing (silently ignoring lists).

Signed-off-by: Uli Schlachter <psychon@znc.in>