Enable the XKB extension
This enables the XKB extension, but skips code generation for the three
requests which currently either cause the code generator to break or to
produce broken Rust code. Everything else seems to at least compile.

This is possible thanks to the many small improvements that happened
recently.

Goes-a-long-way-towards-fixing: https://github.com/psychon/x11rb/issues/56
Signed-off-by: Uli Schlachter <psychon@znc.in>