Allow the clippy cognitive complexity lint
XKB's SetNames request has a <switch> with lots of fields. This causes
the code generator to emit some highly repetitive code for the
implementation of SetNamesAux::to_ne_bytes(). Clippy complains that the
cognitive complexity of this function is too high.

This commit just allows this complex function, because this is
repetitive code, so not as complex as claimed, and this code is not
hand-written, so maintainability is not a concern.

Signed-off-by: Uli Schlachter <psychon@znc.in>