Handle <bitcases> with multiple fields
Before this commit, it was possible to set only some of the fields that
all belong to the same <bitcase> while leaving some others unset. The
result would be a broken request to the server that likely failed to
parse correctly.

This commit fixes that by adding special handling to <bitcase>s with
multiple members: They get their own struct that then provides all
necessary values at once.

Sadly, there are no good names for these bitcases. This commit ends up
with things like "bitcase3" in names. This needs to be improved in the
future, but right now I have no idea how.

Signed-off-by: Uli Schlachter <psychon@znc.in>