Add a "special_cases" python module
This commit adds a new module to the code generator that should handle
"special cases". There are a couple of cases where we do "extra magic"
so that a nicer API is generated. This commit should move this "extra
magic" to an extra, central place. Sadly, I only managed to do so for
two special cases while lots of others still remain (ListFontsWithInfo,
SendEvent).

Signed-off-by: Uli Schlachter <psychon@znc.in>