Simplify interpreting the 'value' in GetPropertyReply
The reply to GetProperty contains a 'value' list. In the XML, this is
described as a list of bytes, i.e. Vec<u8>. However, depending on the
value of the `format` field of the reply, this actually contains u8,
u16, or u32. Interpreting this field correctly was complicated.

This commit adds utility methods to `GetPropertyReply`. value8(),
value16(), and value32() return an iterator that interprets the values
from the reply as u8, u16, and u32, respectively. This iterator stops on
the first parsing error and returns None.

The only possible parsing errors here is "too few bytes in input" and it
should be impossible to hit in well-behaving X11 servers, so I am fine
with interpreting any errors this way.

Fixes: https://github.com/psychon/x11rb/issues/163
Signed-off-by: Uli Schlachter <psychon@znc.in>