Move request handling into its own module
This moves the handling of X11 requests into its own file. The reason
for doing this is file size: After this commit, module.py has 1159 lines
and request.py has 428 lines. Thus, this single function was a huge part
of module.py before.

I only did minimal changes to the code, like name changes and
indentation changes. No "real" changes were done, this is a simple move
of the code.

Signed-off-by: Uli Schlachter <psychon@znc.in>