Implement Serialize for slices
This commit adds an implementation of Serialize for [T] for all T that
implement Serialize and that serialize to something that implements
AsRef<[u8]>. The idea of this latter condition is to accept both arrays
and Vecs.

However, thanks to the LengthAtMost32 situation, this implementation
only works for arrays with "low" lengths. RENDER's Trapezoid type
produces arrays of length 40, which caused some small problems.

The diffstat for this commit is "29 insertions, 21 deletions". So, this
basically moves code from the code generator into Rust (for the new
implementation of Serialize).

The diffstat for the generated code is a bit more substantial: 92
insertions, 351 deletions. So, this gets rid of 259 lines from the
generated code. That's quite nice, plus this makes the generated code
easier to read.

Signed-off-by: Uli Schlachter <psychon@znc.in>