Remove useless type aliases from generated code
The generated code imported GenericEvent and GenericError as
X11GenericEvent and X11GenericError, respectively, to avoid a name clash
with the event "Generic" from the present extension and the error
"Generic" from the GLX extension. Also, "Option" was imported as
"RustOption", because the present extension has an "Option" enum.

This commit gets rid of this renaming and instead changes the generated
code so that it refers to these types in a fully qualified manner where
necessary.

For example, in the present extension, the code now uses
std::option::Option instead of RustOption, while everything else just
uses plain Option.

This should improve the readability of the code a lot. Plus, the renames
appeared in the API docs, which is quite ugly and unfortunate.

Signed-off-by: Uli Schlachter <psychon@znc.in>