Fix types in switch bitcases with multiple values
Previously, each field of a bitcase was handled separately. This meant
that a bitcase containing multiple fields was turned into multiple
fields on the struct, each with type Option<Something>. Now, we get a
single field containing an Option for some new struct. This more closely
matches the on-the-wire representation.

Fixes: https://github.com/psychon/x11rb/issues/160
Signed-off-by: Uli Schlachter <psychon@znc.in>