code generator: Add support for <case>
The code already supported <bitcase>s. These are types where the bit
pattern of some value are examined. The set bits indicate which other
fields are present. These are represented as structs containing
Option<>s.

There are already <case>s. Here, the value itself is compared for
equality. The idea is that only a single match is possible. Thus, this
is better represented as an enum in Rust.

This commit adds the necessary magic to the code generator.

Signed-off-by: Uli Schlachter <psychon@znc.in>