Do not treat popcount references as simple
The code generator contains some heuristics to map a list's length field
to the list. For example, this allows the code for InternAtom to get the
length of the atom name from the given slice.

However, so far this heuristics completely ignores the "kind" of
reference. And this is a problem. The XKB extension contains some
<popcount> references which were also treated this way. However, it is
impossible to "undo" a popcount, so for these cases we need the caller
to provide the correct value for us.

Signed-off-by: Uli Schlachter <psychon@znc.in>