Fix length field calculation with expressions
xproto's SetModifierMapping has two fields:

  <field type="CARD8" name="keycodes_per_modifier"/>
  <list type="KEYCODE" names="keycodes">
    <op op="*">
      <fieldref>keycodes_per_modifier</fieldref>
      <value>8</value>
    </op>
  </list>

So, the length of the list is eight times the given
keycodes_per_modifier value. However, so far the code generator
completely ignored this expression and simply assumed that the length of
the list is keycodes_per_modifier. This commit fixes that.

In the above example, we have:

  length_of_the_list = keycodes_per_modifier * 8

In the generated code, the list is given by the caller as a slice and we
want to compute the value of keycodes_per_modifier so that the caller
does not have to specify it. Thus, we have to transform the above
formula into:

  keycodes_per_modifier = length_of_the_list / 8

This commit does exactly that and also adds an assertion to check that
length_of_the_list is a multiple of 8.

This causes a number of changes to the generated code.

There are seven requests in GLX which are affected, all of which have
lists with double the length of a given value.

In XInput, the SetDeviceModifierMapping request has a list named
"keymaps" whose length is eight times the value of
keycodes_per_modifier.

Similarly, xproto's SetModifierMapping has a list with a length of eight
times the given value.

Fixes: https://github.com/psychon/x11rb/issues/61
Signed-off-by: Uli Schlachter <psychon@znc.in>