RustConnection: Handle connection setup failures
The X11 server can reply to a Setup request not only with a Setup, but
also with two things: SetupFailed and SetupAuthenticate.
SetupAuthenticate is supposed to initiate some challenge-response stuff
for further authentication, but at least libxcb does not implement
anything like this.

Both SetupFailed and SetupAuthenticate contain a reason message. This
commit makes RustConnection return these messages as error that print
these reason messages.

Signed-off-by: Uli Schlachter <psychon@znc.in>