RustConnection: Fix errors for dropped reply cookies
When a cookie for a request which expects a reply is dropped, it informs
the Connection. The connection should then treat any errors that come in
later as events (i.e. return them from wait_for_reply()). However,
RustConnection still internally enqueued the error as a reply to the
request.

Signed-off-by: Uli Schlachter <psychon@znc.in>