RustConnection: Spawn a thread for reading from the server
We may never block in a write, because the server stops reading from us
if its write buffer fills up. libxcb handles this with select() and
reading from the socket even if it tries to write something. Doing the
same thing in Rust does not seem to be too easy, so this commit makes
RustConnection spawn a thread that always tries to read from the
connection. Each X11 packet that is received is stuffed into an mpsc
channel for later handling by RustConnection.

Signed-off-by: Uli Schlachter <psychon@znc.in>