Allow using RustConnection with different Stream
This commit makes RustConnection generic over the stream object, which
is required to implement Read + Write. We set the default stream to be
stream::Stream so that all previous code will still work.

One application of this is that the user can choose if they want to
buffer their stream themselves (regardless of what x11rb decides to do
ultimately).