Refactor xauth handling a bit
The actual reading-of-the-file is moved into a submodule. The outer
module contains the necessary type definitions and handles the matching
of the entries.

Also, this commit adds an iterator over the file entries. The goal here
is to simplify testing of the (to be written) matching code. The idea is
to have an internal function that can be called with an iterator of
entries to match against. The tests can then just call this with another
kind of iterator that does not read entries from a file.

Signed-off-by: Uli Schlachter <psychon@znc.in>