Get the correct address for X auth matching
The previous commits set up the proper code for matching entries from
~/.Xauthority. However, up to now, RustConnection was not specifying
which address is sought. This commit adds this machinery.

rust_connection::stream::Stream gets a new method peer_addr(). This
function returns the address of the peer in exactly the way that is
necessary as input for .Xauthority matching.

For local connections (unix socket, 127.0.0.1 or ::1), this means that
Family::Local with the host name of the computer should be used. To get
the hostname, this commit adds a dependency on the gethostname crate.

For v4 addresses that are not localhost, the octets of the IP address
are used. For v6 addresses, we have to check not only for localhost, but
also for v4-in-v6 addresses. If neither case occurs, the address bytes
to use are again the octets of the IP address.

This was lightly tested with the following commands:

  rm -f /tmp/file
  xauth -f /tmp/file add 192.168.178.41:2 . be3b34142d8b74c95fb6e2a978450810
  Xephyr -listen tcp -auth /tmp/file :2

Now, connecting against this X11 server was attempted with the following
command, with different values for $HOST:

  XAUTHORITY=/tmp/file DISPLAY=$HOST:2 cargo run --example simple_window --no-default-features --features vendor-xcb-proto

Additionally, the code was patched to print out what address was used
for matching and what the resulting auth info was.

With HOST=127.0.0.1, the connection no auth info is found and the X11
server rejects the connection.

With HOST=192.168.179.41 (my local IP), the right auth info is found and
the connection succeeds.

With HOST=::1, the correct address is found (host name of this PC), but
no auth info matches.

Thus, I believe that for now this code works correctly, but of course it
would be nice to have more (automated) tests for this. However, I am now
tired of this X11 auth stuff and testing this new peer_addr() in a unit
test seems complicated (due to the necessary network use).

Fixes: https://github.com/psychon/x11rb/issues/196
Signed-off-by: Uli Schlachter <psychon@znc.in>